refactor(video): keep shared contracts in core type leaf
